% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Depth_Outlier_Funcs.R
\name{bootstrap_C.alpha}
\alias{bootstrap_C.alpha}
\title{Alpha method to bootstrap a cutoff value to identify anomalies}
\usage{
bootstrap_C.alpha(coeff, d, B, alpha, d.method)
}
\arguments{
\item{coeff}{A dataframe of coefficients of interest. Each column is for the parameter to be estimate. Each row is the estimated parameters
fore each curve. (No ID column here)}

\item{d}{A vector of depths for all of the coefficients.}

\item{B}{A value determining how many bootstrap datasets should be made to estimate
the cutoff value with a suggested rate of 1000.}

\item{alpha}{A value determining the percentage of rows to remove from \code{coeff}.
\code{alpha} should be between (0, 1) with a suggested value of 0.05. Do not need to
identify if \code{c.method} = "depth".}

\item{d.method}{A character string determining the depth function to use: "LP", "Projection",
"Mahalanobis", or "Euclidean". It is suggested to not use "Tukey" due to singularity in
coefficient matrix. For details see \code{\link[DepthProc]{depth}}}
}
\value{
\code{$Cb} the cutoff value
}
\description{
Alpha method to bootstrap a cutoff value to identify anomalies
}
\details{
The "alpha" \code{c.method} removes the alpha percent least deep coefficients. The rest of the
coefficients are bootstrapped and new depths are computed for each new bootstrapped set. The
1% empirical percentile of the depths is saved. The cuttoff value is the median of these
1% empirical percentile of the depths.
}
\seealso{
\code{\link[DepthProc]{depth}}, \code{\link{bootstrap_C.alpha}}, and
 \code{\link{bootstrap_C.depth}}
}
