% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Knot_Selection_Funcs.R
\name{choose_knots}
\alias{choose_knots}
\title{Choose knots over data with unequal lengths}
\usage{
choose_knots(Y, K)
}
\arguments{
\item{Y}{Matrix of functional datasets. Should contain \code{$ID}
indicating the separate curves, \code{$x} the time points, and \code{y} the observations.}

\item{K}{Total number of knots.}
}
\value{
The knots based on differing endpoints.
}
\description{
Choose knots over data with unequal lengths
}
\details{
\code{x} should be the predictor variable for the longest dataset in \code{y}.
Where there are observations in \code{y} the predictor variable should be the same as \code{x}.
}
