% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Depth_Outlier_Funcs.R
\name{depth_Outliers}
\alias{depth_Outliers}
\title{Identify anomalous coefficients using depth}
\usage{
depth_Outliers(coeff, d.method = "L2", c.method = "depth", alpha = 0.05,
  B = 1000)
}
\arguments{
\item{coeff}{A dataframe of coefficients of interest. The first column is \code{ID} identifier.
The rest of the columns are for the parameter to be estimate. Each row is the estimated parameters
fore each curve.}

\item{d.method}{A character string determining the depth function to use: "LP", "Projection",
"Mahalanobis", or "Euclidean". It is suggested to not use "Tukey" due to singularity in
coefficient matrix. For details see \code{\link[DepthProc]{depth}}}

\item{c.method}{A character string determining the method to estimate the cutoff value.
This can be "depth" or "alpha".}

\item{alpha}{A value determining the percentage of rows to remove from \code{coeff}.
\code{alpha} should be between (0, 1) with a suggested value of 0.05. Do not need to
identify if \code{c.method} = "depth".}

\item{B}{A value determining how many bootstrap datasets should be made to estimate
the cutoff value with a suggested rate of 1000.}
}
\value{
A list with \code{$outliers} is a tibble containing the ID, outlier identification, depth value.
True identifies outliers. \code{$Cb} is the cutoff value
}
\description{
Identify anomalous coefficients using depth
}
\details{
The function uses a bootstrap method to estimate a cutoff depth value.
Depths below this cutoff depth value are flagged as anomalous.

The "alpha" \code{c.method} removes the alpha percent least deep coefficients. The rest of the
coefficients are bootstrapped and new depths are computed for each new bootstrapped set. The
1% empirical percentile of the depths is saved. The cuttoff value is the median of these
1% empirical percentile of the depths.

The "depth" \code{c.method} bootstraps the coefficients with probability related to the
original depth values.  New depths are computed for each new bootstrapped set. The
1 percent empirical percentile of the depths is saved. The cuttoff value is the median of these
1 percent empirical percentile of the depths.
}
\references{
\insertRef{febrero2008}{unequalgroupoutlier}
}
\seealso{
\code{\link[DepthProc]{depth}}, \code{\link{bootstrap_C.alpha}}, and
 \code{\link{bootstrap_C.depth}}
}
