% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mixed_Model_Funcs.R
\name{Mixed_Spline_Fit_Single}
\alias{Mixed_Spline_Fit_Single}
\title{Fitting single dataset with mixed models and splines}
\usage{
Mixed_Spline_Fit_Single(x, y, model)
}
\arguments{
\item{x}{A vector of a predictor variable.}

\item{y}{A vector of dependent variable}

\item{model}{A list containing \code{Z_p}, the matrix formed from a spline basis,
and \code{X_p}, the matrix formed to fit the trend. \code{Z_p} is associated
with the random effects and \code{X_p} is associated with the fixed effects.}
}
\value{
\code{fit} is the vector of fitted values.
\code{coeff_fix} is the vector of fixed effects coefficients.
\code{coeff_rand} is the vector of random effects coefficients.
}
\description{
This function fits a mixed model.
}
\details{
Must provide a model.
}
\examples{
knots <- seq(1, 1000, len = 40)
x <- c(1:1000)
y <- rnorm(1000, x, 25)
model <- Mixed_Spline_Model(x, knots, P = 1)
fit <- Mixed_Splinefit(x, y , model)
plot(fit$fit, type = "l")
}
\references{
\insertRef{wand2003}{unequalgroupoutlier}
}
\seealso{
\code{\link{Mixed_Spline_Model}} for how to create a model.
}
