% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mixed_Model_Funcs.R
\name{Mixed_Spline_Model}
\alias{Mixed_Spline_Model}
\title{Create matrices for the mixed model with splines}
\usage{
Mixed_Spline_Model(x, knots, P = 1, spline = "Radial", r = 1, theta = 2)
}
\arguments{
\item{x}{A vector of a predictor variable.}

\item{knots}{A vector containing the knots for the basis.}

\item{P}{OPTIONAL: A value specifying the degree of the polynomial.}

\item{spline}{OPTIONAL: The basis to be used for the spline model.
This can either be "Truncated Poly" for the truncated polynomial
basis function or "Radial" for the radial basis function.}

\item{r}{OPTIONAL: A value for the degree of the radial basis function. Not
used in the truncated polynomial basis function.}

\item{theta}{OPTIONAL: A value for the range of the radial basis function. Not
used in the truncated polynomial basis function.}
}
\value{
\code{Z_p} is a basis matrix dimensions = n x k
where n equals the length of \code{x} and k equals the length of \code{int_knots}.
This matrix is used for the random part of the mixed model. \code{X_p} is the matrix
with dimensions n x P. The columns are: the intercept, x, x^2, ..., x^P. This models the mean function of
the data. \code{X_p} is used for the fixed part of the mixed model.
}
\description{
This function creates the two matrices for the fixed and
random components of the mixed model. The matrix corresponding to the
random components is a made from a spline basis.
}
\details{
The truncated polynomial basis function is of the form: (x - k)+^P.
The radial basis function is of the form exp(-abs(x - k)^r / theta).#'
Set P = 1 for linear trend.
}
\examples{
knots <- seq(1, 1000, len = 40)
x <- c(1:1000)
y <- rnorm(1000, x, 25)
model <- Mixed_Spline_Model(x, knots, P = 1)
}
\references{
\insertRef{wand2003}{unequalgroupoutlier}
}
\seealso{
\code{\link{radial_basis_fun}} and \code{\link{piecewise_poly_basis_fun}}
}
