% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basis_Funcs.R
\name{piecewise_poly_basis_fun}
\alias{piecewise_poly_basis_fun}
\title{Truncated Piecewise polynomial basis}
\usage{
piecewise_poly_basis_fun(x, int_knots, p)
}
\arguments{
\item{x}{A vector of a predictor variable.}

\item{int_knots}{A vector containing the interior knots.}

\item{p}{A value specifying the degree of the polynomial.}
}
\value{
A matrix consisting of (x_i - k_j)^p if x_i > k_j or 0 otherwise for
i = 1,...,n and j = 1,... k. Where n equals the length of \code{x} and k equals the length of \code{int_knots}.
}
\description{
Truncated Piecewise polynomial basis
}
\examples{
out <- piecwise_poly_basis_fun(x = c(1:9), k = c(2:4), p = 3)
plot(out[,1], type = "l") # see what your basis looks like

}
\references{
\insertRef{wand2003}{unequalgroupoutlier}
}
