% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basis_Funcs.R
\name{radial_basis_fun}
\alias{radial_basis_fun}
\title{Radial basis}
\usage{
radial_basis_fun(x, int_knots, p, r = 1, theta = 2)
}
\arguments{
\item{x}{A vector of a predictor variable.}

\item{int_knots}{A vector containing the interior knots.}

\item{p}{A value specifying the degree of the polynomial.}

\item{theta}{OPTIONAL: A value specifying the range parameter.}
}
\value{
\code{Z} A radial basis matrix dimensions = n x k where n equals the length of \code{x} and k equals the length of \code{int_knots}.
}
\description{
Radial basis
}
\details{
Let the {i, j} value of \code{Z_p} be exp( - abs(x_i - k_j)^p / theta) and the {i, j}
 value of \code{Omega} be exp( - abs(k_i - k_j)^p / theta), then \code{Z} = \code{Z_p}%%*%%\code{Omega}^-1/2. The correction
 is required to be able to use this matrix in the mixed model framework.
}
\examples{
radial_basis_fun(x = c(1:9), k = c(2:4), p = 3)
}
\references{
\insertRef{wand2003}{unequalgroupoutlier}
}
