% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateToyData.R
\name{Unequal_ToyData}
\alias{Unequal_ToyData}
\title{Create Unequal functional data}
\usage{
Unequal_ToyData(M = 100, A = 10, N = 1000)
}
\arguments{
\item{M}{Number of non-anomalous functional datasets}

\item{A}{Total number of anomalous datasets}

\item{N}{Number of total possible observations in each dataset.}
}
\value{
A list containing the toy data \code{data} each row is a dataset.
\code{label} for if the dataset is normal ("N") or anomalous ("A1" or "A2")
Three functions used to create the functional data: \code{f_x}, the function used to create the normal data,
\code{f_x.A1} the function used to create one set of anomalous data, and \code{f_x.A2} the
function used to create the second set of anomalous data.
}
\description{
Create Unequal functional data
}
